1. Big refactoring - removed ELogSystem facade, now using global extern functions. 2. Fixed more pedantic compiler warnings 3. Disabled some warnings when including protoc generated headers (message and gRPC service headers) 4. Using flat gRPC array instead of vector due to some warnings about move ctor (not required in any way). 5. Fixed/checked more signed/unsigned warnings 6. Disabled some warnings when including mysql jdbc driver headers 7. Set default deadline value for gRPC log target (was missing, used -1, now set to 60 seconds, if value too small synchronous streaming client tests fail) 8. Moved some gRPC base reactor call data members to inline file, so that ELOG_INVALID_REQUEST_ID can be used 9. Added more trace messages to debug the base reactor state machine and pending queue state 10. Fixed SEVERE BUG in gRPC base reactor - when done writing one message and another one is pulled from the pending queue - then the in-flight request id is not updated. The result is that during next round of OnWriteDone(), the call data object cannot be retrieved (because the request id was not updated), and cannot be cleaned up. The result is resource leak. In best case client will get stuck. In worst case two in-flight messages may use the same call data. The current code shows that all logging threads will get stuck in a busy loop. NOW BUG IS FIXED. Upon OnWriteDone(), if another message is pulled from the pending queue, then the in-flight request id is updated properly, so next round of OnWriteDone() can cleanup the call data entry. 11. FIxed assert being fired in gRPC base reactor due to missing reset of in-flight flag to false (now being set to false when pending request queue is empty, as it should, so another assert was also added for the empty pending request queue check). 12. FIxed bug in optional parameter loading (config loader) - not checking if found on some scenarios, and so returned value is zero (overriding previous user default, when it shouldn't, because configuration item is not found) 13. Added getLogTargetUInt32Property() and getOptionalLogTargetInt32Property() to ELogConfigLoader 14. Fixed GRPC tests to really test message arrive to the server. Also unified all tests into one template function.

+++ src/elog/src/elog_datadog_target.cpp
@@ -50,7 +50,7 @@ bool ELogDatadogTarget::stopLogTarget() { return m_client.stop(); }
 
 uint32_t ELogDatadogTarget::writeLogRecord(const ELogRecord& logRecord) {
     ELOG_REPORT_TRACE("Preapring log message for Datadog");
-    uint32_t index = m_logItemArray.size();
+    size_t index = m_logItemArray.size();
 
     // log line
     std::string logMsg;
@@ -99,7 +99,7 @@ uint32_t ELogDatadogTarget::writeLogRecord(const ELogRecord& logRecord) {
     }*/
 
     ELOG_REPORT_TRACE("Log message for Datadog is ready, body: %s", m_logItemArray.dump().c_str());
-    return logMsg.size();
+    return (uint32_t)logMsg.size();
 }
 
 void ELogDatadogTarget::flushLogTarget() {
